refactor: refactor generate command
