+++ pkg/controller/generate.go
@@ -21,31 +21,50 @@ type FindingPackage struct {
 	RegistryName string
 }
 
-func (ctrl *Controller) Generate(ctx context.Context, param *Param, args ...string) error { //nolint:cyclop,funlen
+// Generate searches packages in registries and outputs the configuration to standard output.
+// If no package is specified, the interactive fuzzy finder is launched.
+// If the package supports, the latest version is gotten by GitHub API.
+func (ctrl *Controller) Generate(ctx context.Context, param *Param, args ...string) error {
+	list, err := ctrl.generate(ctx, param, args...)
+	if err != nil {
+		return err
+	}
+	if list == nil {
+		return nil
+	}
+	if err := yaml.NewEncoder(ctrl.Stdout).Encode(list); err != nil {
+		return fmt.Errorf("output generated package configuration: %w", err)
+	}
+	return nil
+}
+
+func (ctrl *Controller) generate(ctx context.Context, param *Param, args ...string) (interface{}, error) { //nolint:cyclop
 	wd, err := os.Getwd()
 	if err != nil {
-		return fmt.Errorf("get the current directory: %w", err)
+		return nil, fmt.Errorf("get the current directory: %w", err)
 	}
+
 	cfgFilePath := ctrl.getConfigFilePath(wd, param.ConfigFilePath)
 	if cfgFilePath == "" {
-		return errConfigFileNotFound
+		return nil, errConfigFileNotFound
 	}
 	cfg := &Config{}
 	if err := ctrl.readConfig(cfgFilePath, cfg); err != nil {
-		return err
+		return nil, err
 	}
 	if err := validateConfig(cfg); err != nil {
-		return fmt.Errorf("configuration is invalid: %w", err)
+		return nil, fmt.Errorf("configuration is invalid: %w", err)
 	}
 	registryContents, err := ctrl.installRegistries(ctx, cfg, cfgFilePath)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	if param.File != "" || len(args) != 0 {
 		return ctrl.outputListedPkgs(ctx, param, registryContents, args...)
 	}
 
+	// maps the package and the registry
 	var pkgs []*FindingPackage
 	for registryName, registryContent := range registryContents {
 		for _, pkg := range registryContent.PackageInfos {
@@ -55,59 +74,18 @@ func (ctrl *Controller) Generate(ctx context.Context, param *Param, args ...stri
 			})
 		}
 	}
-	idx, err := fuzzyfinder.Find(pkgs, func(i int) string {
-		pkg := pkgs[i]
-		files := pkg.PackageInfo.GetFiles()
-		fileNames := make([]string, len(files))
-		for i, file := range files {
-			fileNames[i] = file.Name
-		}
-		fileNamesStr := strings.Join(fileNames, ", ")
-		pkgName := pkg.PackageInfo.GetName()
-		if strings.HasSuffix(pkgName, "/"+fileNamesStr) || pkgName == fileNamesStr {
-			return fmt.Sprintf("%s (%s)", pkgName, pkg.RegistryName)
-		}
-		return fmt.Sprintf("%s (%s) (%s)", pkgName, pkg.RegistryName, fileNamesStr)
-	},
-		fuzzyfinder.WithPreviewWindow(func(i, w, h int) string {
-			if i < 0 {
-				return ""
-			}
-			pkg := pkgs[i]
-			return fmt.Sprintf("%s\n\n%s\n%s",
-				pkg.PackageInfo.GetName(),
-				pkg.PackageInfo.GetLink(),
-				formatDescription(pkg.PackageInfo.GetDescription(), w/2-8)) //nolint:gomnd
-		}))
+
+	// Launch the fuzzy finder
+	idx, err := ctrl.launchFuzzyFinder(pkgs)
 	if err != nil {
 		if errors.Is(err, fuzzyfinder.ErrAbort) {
-			return nil
+			return nil, nil //nolint:nilnil
 		}
-		return fmt.Errorf("find the package: %w", err)
+		return nil, fmt.Errorf("find the package: %w", err)
 	}
 	pkg := pkgs[idx]
-	outputPkg := ctrl.getOutputtedPkg(ctx, pkg)
-	if err := yaml.NewEncoder(ctrl.Stdout).Encode([]interface{}{outputPkg}); err != nil {
-		return fmt.Errorf("output generated package configuration: %w", err)
-	}
-	return nil
-}
 
-func formatDescription(desc string, w int) string {
-	descRune := []rune(desc)
-	lenDescRune := len(descRune)
-	lineWidth := w - len([]rune("\n"))
-	numOfLines := (lenDescRune / lineWidth) + 1
-	descArr := make([]string, numOfLines)
-	for i := 0; i < numOfLines; i++ {
-		start := i * lineWidth
-		end := start + lineWidth
-		if i == numOfLines-1 {
-			end = lenDescRune
-		}
-		descArr[i] = string(descRune[start:end])
-	}
-	return strings.Join(descArr, "\n")
+	return []interface{}{ctrl.getOutputtedPkg(ctx, pkg)}, nil
 }
 
 func getGeneratePkg(s string) string {
@@ -117,7 +95,7 @@ func getGeneratePkg(s string) string {
 	return s
 }
 
-func (ctrl *Controller) outputListedPkgs(ctx context.Context, param *Param, registryContents map[string]*RegistryContent, pkgNames ...string) error { //nolint:cyclop
+func (ctrl *Controller) outputListedPkgs(ctx context.Context, param *Param, registryContents map[string]*RegistryContent, pkgNames ...string) (interface{}, error) {
 	m := map[string]*FindingPackage{}
 	for registryName, registryContent := range registryContents {
 		for _, pkg := range registryContent.PackageInfos {
@@ -133,43 +111,48 @@ func (ctrl *Controller) outputListedPkgs(ctx context.Context, param *Param, regi
 		pkgName = getGeneratePkg(pkgName)
 		findingPkg, ok := m[pkgName]
 		if !ok {
-			return logerr.WithFields(errUnknownPkg, logrus.Fields{"package_name": pkgName}) //nolint:wrapcheck
+			return nil, logerr.WithFields(errUnknownPkg, logrus.Fields{"package_name": pkgName}) //nolint:wrapcheck
 		}
 		outputPkg := ctrl.getOutputtedPkg(ctx, findingPkg)
 		outputPkgs = append(outputPkgs, outputPkg)
 	}
 
-	if param.File != "" { //nolint:nestif
-		var file io.Reader
-		if param.File == "-" {
-			file = ctrl.Stdin
-		} else {
-			f, err := os.Open(param.File)
-			if err != nil {
-				return fmt.Errorf("open the package list file: %w", err)
-			}
-			defer f.Close()
-			file = f
+	if param.File != "" {
+		pkgs, err := ctrl.readGeneratedPkgsFromFile(ctx, param, outputPkgs, m)
+		if err != nil {
+			return nil, err
 		}
-		scanner := bufio.NewScanner(file)
-		for scanner.Scan() {
-			txt := getGeneratePkg(scanner.Text())
-			findingPkg, ok := m[txt]
-			if !ok {
-				return logerr.WithFields(errUnknownPkg, logrus.Fields{"package_name": txt}) //nolint:wrapcheck
-			}
-			outputPkg := ctrl.getOutputtedPkg(ctx, findingPkg)
-			outputPkgs = append(outputPkgs, outputPkg)
-		}
-		if err := scanner.Err(); err != nil {
-			return fmt.Errorf("failed to read the file: %w", err)
-		}
+		outputPkgs = pkgs
 	}
+	return outputPkgs, nil
+}
 
-	if err := yaml.NewEncoder(ctrl.Stdout).Encode(outputPkgs); err != nil {
-		return fmt.Errorf("output generated package configuration: %w", err)
+func (ctrl *Controller) readGeneratedPkgsFromFile(ctx context.Context, param *Param, outputPkgs []*Package, m map[string]*FindingPackage) ([]*Package, error) {
+	var file io.Reader
+	if param.File == "-" {
+		file = ctrl.Stdin
+	} else {
+		f, err := os.Open(param.File)
+		if err != nil {
+			return nil, fmt.Errorf("open the package list file: %w", err)
+		}
+		defer f.Close()
+		file = f
 	}
-	return nil
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		txt := getGeneratePkg(scanner.Text())
+		findingPkg, ok := m[txt]
+		if !ok {
+			return nil, logerr.WithFields(errUnknownPkg, logrus.Fields{"package_name": txt}) //nolint:wrapcheck
+		}
+		outputPkg := ctrl.getOutputtedPkg(ctx, findingPkg)
+		outputPkgs = append(outputPkgs, outputPkg)
+	}
+	if err := scanner.Err(); err != nil {
+		return nil, fmt.Errorf("failed to read the file: %w", err)
+	}
+	return outputPkgs, nil
 }
 
 func (ctrl *Controller) listAndGetTagName(ctx context.Context, pkgInfo *PackageInfo) string {
